TS-3 fix for finding TCL library
This patch adds 'm4/tcl.m4' file taken from tcl 8.5.7:
http://prdownloads.sourceforge.net/tcl/tcl8.5.7-src.tar.gz

And includes the macros per automake convention:
http://www.gnu.org/software/automake/manual/automake.html#Local-Macros

I've removed AC_PATH_PROG for tclsh since it is not used anywhere in
the build.

I'm also assuming if a valid tclConfig.sh is found and can load properly,                                                       
then the libraries are built properly and working. This is because
tclConfig.sh is a generated file from a successful build of the tcl
libraries.

Also taking the time to svn:ignore aclocal.m4 since it is generated
by configure.


git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/trunk@833453 13f79535-47bb-0310-9956-ffa450edef68