TS-307 Add support for a dedicate DNS thread.
This is disabled by default, but can be enabled with

     CONFIG proxy.config.dns.dedicated_thread INT 1

I'd imagine this is primarily useful on systems with either plenty of
physical CPUs to spare, or where there are a lot of DNS lookups (e.g.
transparent proxy quite likely). Enabling this will avoid uneven load
on the #0 net-thread.

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1027765 13f79535-47bb-0310-9956-ffa450edef68