TS-533 and TS-135
This is a sad day of defeat.  Not my defeat, but more a collective human defeat.

Question: "Chips fabricated today don't have 64bit atomic primitives?"
Answer: "Be sad."

The ARM box we're working on (armv5tejl) doesn't support any 64bit primitives.

This means we need a method of using a global (yes, giant lock of death) to
protect modifications of arbitrary 64bit integers in process space.  We could
make this less contentious by allocating pagesize/8 mutexs and then protecting
an int64 based on its page offset.  Instead, I think we should mobilize to
burn these architectures to the ground and use public embarrassment to fix
future instruction sets.

If another platform has this issue, we'll want to change the define to:
TS_ARCHITECTURE_LACKS_64BIT_INSTRUCTIONS and turn on the global death lock
based on that.

This does not change performance on any other platform -- it's compile time
capital punishment.


git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1034786 13f79535-47bb-0310-9956-ffa450edef68