TS-18 bring IPv6 functionality back to trunk
Tested: OSX-10.6.5

This commit reverts r959407 r959405 r959404 and brings in the fix
on the 'ipv6' branch for Mac OSX:

Reapply r953510 "added IPv6 support for incoming requests". This
is the first step in bringing back the ipv6 changes to trunk. Most
of the conflicts resulted in the transparent proxy changes, but I
believe I've been able to resolve them without issue. This will
definitely need some testing to verify I have not broke anything
in the process.

Reapply r953892 "added stats for ipv4 and ipv6 request via REC".
No conflicts, easy merge.

Reapply r958062 "added INKHttpTxnClientSockAddrGet()". This is a
small patch with a few easy-to-resolve conflicts as a result of the
renaming of the InkAPI stuff like TS_SUCCESS.

Merge in r997943 from ipv6 branch. This is the fix that allows IPv6
to work properly on Mac OSX.

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1040017 13f79535-47bb-0310-9956-ffa450edef68