Some final tweaks to resolve TS-402.
TsConfig now has better automake integration. The lexical and grammar files
are built using implicit rules so all of the dependecies are handled correctly.
Additional code changes eliminated the need for a lexical header file and
the circular dependency betwee the lexical and grammar files.

WCCP was updated to support the REMOVAL_QUERY message. It was also updated to
check the assignment information reported by routers and resend the cache
assignment if the router's view is not consistent with the cache view. This
seems to resolve a timing issue with restarting Traffic Server.



git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1044101 13f79535-47bb-0310-9956-ffa450edef68