TS-638 Eliminated CLI (R.I.P.)
This patch eliminates the rest of mgmt/cli. It also extends the cli2
APIs with a new public API:

     inkapi INKError INKBounce(bool cluster);

This works similar to INKRestart(), except it will only restart
the traffic_server process (it will not touch traffic_manager).

Finally, this patch eliminates the need for using the Layout engine
in a client app. If you pass in NULL as the socket path, INKInit()
will figure it out for you.

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1061011 13f79535-47bb-0310-9956-ffa450edef68