TS-649 Dynamic libraries for mgmt
This makes libts -> libtsutil, and we now support
making .so's for libtsutil.so and libtsmgmt.so. All
binaries are changed to use this, except traffic_server
which continues to use the libtsutil.a library (for
performance on e.g. 32-bit platforms).

This also renames the public API include file to be

     #include <ts/mgmtapi.h>

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1061490 13f79535-47bb-0310-9956-ffa450edef68