TS-609 Added suport for treating the 64-bit config ints as 8-bit ints
This is particularly useful in the Http Transaction class, because
a larger portion of those configs are now overridable per transaction,
so this saves a chunk of memory on every active transaction.

It might make sense to modify other core modules to use the new
MgmtByte as well, but it's probably not critical. Any config that is
a "toggle" (on|off), or small integers (0..3) are candidates to use
this new type.

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1079662 13f79535-47bb-0310-9956-ffa450edef68