[TS-845] make proxy.config.cluster.ethernet_interface default to lo/lo0
lo on linux and lo0 on bsd derivatives. it does no effect when you
really want to setup the cluster type=1, in that case you will need
to change the interface to a real interface like eth0. this patch
will just make it is safe to remove the whole cluster config block
from records.config if you like, and don't drive the newbie crazy
when they copy the records.config to another system where there is
no eth0 but bonding0.
we may need to rework the whole ClusterComunication system initialize
when we have someone cleanup the mgmt/web2(the webui).

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1137894 13f79535-47bb-0310-9956-ffa450edef68