TS-875 TSFetchRestpGet(), TSFetchPageResptGet() and TSFetchUrl() have incorrect asserts
Author: Manjesh Nilange

At some point (up to you Manjesh) we might want to file an RFE to add
a data type for this, and change these (deprecate and make new) APIs.

We've already added a number of typedef's to make the type checking
better, so it'd make sense to do so for this API too.

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1146414 13f79535-47bb-0310-9956-ffa450edef68