TS-940
On bleeding-edge Illumos, this can be set on the fly. And, if the Linux
implementation is added, this flow makes more sense.  The initial congestion
window is set on the parent socket (that which is bound) and then it
can be set again immediately before the first write.  This allows you to
set it to 10 in records.conf and then use a module to set it back down to
something like 3 or 4 for mobile devices.  I used conf_remap to test
on-the-fly alteration of the congestion window.


git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1164676 13f79535-47bb-0310-9956-ffa450edef68