TS-1135: index wildcard certificates for ServerNameIndication
Add a SSLContextStorage support class. This objects owns all the
SSL contexts and indexes them by both name and (stringified) address.
Both the CN from the certificate subject name and the DNS
subjectAlternativeNames are indexed.

If we find a wildcard name, we use the trie for indexing so that
we get longest match semantics. We only support wildcards of the
form '*.foo.org' because that's the only useful form, and the code
for supporting the general case would be a bit hairy.