TS-1147: Implement default certificate fallback.
Since we removed the certificate specification from records.config,
implement a new default fallback mechanism in ssl_multicert.config.
We can now use this default when the client does not give us a
hostname, or when the the hostname lookup fails.

Update the ssl_multicert.config parser to accept lines that don't
have a dest_ip field. Update the documentation in ssl_multicert.config.