TS-1322: CONNECT to parent proxy has URL with a trailing slash
I want to redirect a CONNECT request to parent proxy.  I'm doing
it with the TSHttpTxnParentProxySet API.  The problem is that the
URL that the TS is setting is invalid:

  CONNECT www.example.com:443/ HTTP/1.1
  Host: www.example.com:443

The TS adding to every URL a trailing slash in the end of the URL,
but here it makes the URL invalid (look at
http://www.ietf.org/rfc/rfc2817.txt in section 5.2).

I'm suggesting a patch for the URL.cc file that doesn't add a
trailing slash when not needed (when there is only hostname).  This
is valid action as is written in the RFC
(http://tools.ietf.org/html/rfc3986 in section 3) : "When authority
is present, the path must either be empty or begin with a slash
("/") character" While" authority" is the hostaname.