TS-1249: preserve gzip encoding in ESI plugin
These lines are removing "Accept-Encoding" from the requests that
the ESI plugin sent based on ESI include in the ESI response.
Essentially they are stopping the ESI include from using gzipped
for the response. Commenting this out give this choice back to user.
User can choose whether to have Accept-Encoding header or not in
his original request and the ESI include will also be fired with
the header if present.