TS-1249: Disable ESI packed nodes by default
The ESI plugin has an internal mechanism that can try to cache a
more efficient version (called packed node in the code) of the ESI
response instead of the raw text. The hope, I think, is to make
parsing much easier for subsequent requests. The mechanism works
like this:
  a) Determine that the ESI response has the proper cache control
     header and is cacheable
  b) Issue a internal "POST" request to the same URL with the same
     requests headers. The post data is the packed node
     version of the ESI response
  c) Intercept this internal request and replace the existing cache
     with it.

I find this mechanism extremely hard to get to work. So I turn if
off completely. It is also later on found to be a incorrect mechanism,
too, in yahoo!. The mechanism does not take into account that key
for HTTP cache can be changed with the Vary header in the response.