Log.cc: Acquire flush_mutex before using it in ink_cond_wait, or we risk unknown behavior.
On FreeBSD/OpenBSD, ink_cond_wait requires the mutex to be locked prior to being callled,
or it will fail with EPERM and cause an infinite loop with no waiting time (i.e. 100% cpu).
This should in theory also apply to Linux, at least according to documentation, so adding an
acquisition of the mutex before the wait should apply to all OSes. This also, in part, solves
issues 1 through 3 described in TS-993.