Make TS build on FreeBSD with hwloc installed
Currently, building TS on FreeBSD and Solaris fails, as cpu_set_t as well as
CPU_SET and CPU_ZERO are not defined within the usual includes. This is
fixed by adding a clause, which includes sys/cpuset.h as well as
pthread_np.h (for pthread_setaffinity_np). I do not have access to a Solaris
machine myself, so I won't know if this works till the buildbot starts working
on it.

Also, this fixes a small issue with a thread pointer being referenced as %lu
instead of the correct %p.