TS-1607: Decouple SSL context indexing from the rest of iocore
Decouple SSL context indexing from the rest of iocore so that we
can make a standalone SSL lookup test harness.

Add a certlookup test program that tests SSL certificate lookup.
This also ingests the top-N web site CVS from Alexa so that we can
test memory usage and lookup performance.

Move all the stand-alone SSL utilities into SSLUtils.cc. Refactor
this a bit to make the code simpler and cleaner.

Teach SSLCertLookup to insert and lookup of IpEndpoint objects.

Using the SSLConfig pattern, decouple the SSLCertLookup from the
online reconfiguration.

Keep a Vec<SSL_CTX> of the contexts that we index so that it's easy
to release them when we reload the contexts. This also sames us
from leaking memory if we have multiple certificates for the same
name. Add memory leak checking to the tests (Darwin only).