TS-1550: remove unnecessary USE_CONFIG_PROCESSOR define
Remove USE_CONFIG_PROCESSOR and enable the use of ProxyConfig for
SSL configuration. Rename classes with a Ssl prefix to SSL to be
consistent with the rest of the SSL naming conventions. Remove
unnecessary global variable sslTerminationConfig. Remove the
SSLConfigParams member from SSLCertLookup because we don't have a
refcount on it to guarantee its lifetime.