vote, fix structure, comment
There's a reason for this silly structure: It's really
convinient to work with when doing backports and only
having to do minimal manipulations, to the output of:

   sed -ne '/PATCHES ACCEPTED TO BACKPORT FROM TRUNK/,/PATCHES PROPOSED TO BACKPORT FROM TRUNK/p' STATUS

Please stick to it, or fix it yourself when you notice
it's gone off. (Ideally, at some point, I'll be replaced
with a very small script that does backports on its own
without my intervention. The subversion project already
does this:
https://svn.apache.org/repos/asf/subversion/trunk/tools/dev/svn-merge-revs.py