TS-977 RecCore: refine P_RecCore.i and rename it to P_RecCore.cc
The suffix of P_RecCore.i is weird for us:
1) cscope can't find the function definition in it by default.
2) using *.i suffix usually indicates that there are something
   bad in design. We can share the code gracefully if spliting the
   function/class carefully.

Let's refine the code so that we can rename it to P_RecCore.cc safely.
BTW, remove an useless file: ./iocore/utils/diags.i.

I have noticed that there are several test_*.i files which is not so
important for us, just keep it as it is.

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>
Signed-off-by: Zhao Yongming <ming.zym@gmail.com>