TS-1585: plugins need memory barriers for ARM
In lulu.h barriers are defined only for x86 and x86_64, but not for
arm.  Well, thankfully implementing barriers for arm was rather
easy (at least, the tests pass here).

Fix ARM build failures by adding ARM barrier intrinsics which
actually do something. This fixes build failues on the ARM port
after promoting the header filter module to main. The patch
denoted below originally was contributed by a Debian porter.
The original issue was discussed in TS-1473.
A no-op fix was proposed to be backported and committed as
TS-1498. However, later TS-1560 was committed to trunk as an
improved version of the same problem.

test/review: igalic, zwoop, arno
backport: igalic