First attempt at creating our docs using sphinx
My intent is to use sphinx to create both, the admin and the sdk docs.
I will transform them from Markdown, ditching the CMS (except, perhaps
for the site itself). The reasoning behind this is so that we have the
docs in-tree, making it more likely for developers to actually
contribute to them.

I am also planning to use the sphinx-doxygen bridge[1] where the CPP
domain[2] does not suffice. or rather, where we need explicit linking
to otherwise duplicated documentation in the doxygen.

[1] http://michaeljones.github.com/breathe/
[2] http://sphinx.readthedocs.org/en/latest/domains.html