TS-1058: add TSHttpTxnCloseAfterResponse experimental API
Add a new API to force the server end of a HTTP keepalive. This is
useful for a plugin that sets the error body to easuly intercept a
transaction, but needs to ensure that the subsequent request does
not go to a different replica. One example of this use case is a
challenge-response authentication protocol where the challenge data
is held in memory on the session.