TS-1735: Remove dead code that invokes missing vmap_config tool
Virtual IPs were once managed such that within a cluster they would
automatically rebalance themselves between nodes by bringing
subinterfaces up and down. After ATS was open sourced the original
setuid tool, vip_config (or traffic_vip_config) was inadvertently
removed; but the code which depended on this tool was not cleaned
up.

Since modern deployments either do not use this tool at all (because
it was broken for a few years) and modern deployments also have
some central system for managing cluster state reliably, we do not
need VMap to implement some scheme for automatically rebalancing
the ips.

This patch keeps much of the code for detecting ip address conflicts
and for receiving the multicast messages from the cluster; but we
remove all instances where we either bring up/down an interface.
Deployments should manage this through external state systems.

Note: VIPs do not actually bind to the specific addresses in vaddrs;
this is just an operations convience to ensure that a cluster has
no ip conflicts or unmanaged vips. This feature becomes even less
useful. LocalManager.cc would have to be modified in some way to
set this up properly.