TS-1772: Remove multiple TS_INLINE defines
It took me a couple of tries to understand what the Inline.cc files
were for, but it's actually not necessary for TS_INLINE to be defined
in many different places. All we need to do to get correct inlining
is define an extern version, and make an inline version available
in a header. Forward declarations should be extern, not TS_INLINE.