TS-1889: refactor remap plugin request URL handling
For (apparantly) historical reasons, the remap plugin support
duplicated the request URL rewriting code. There's really no need for
this, and it's actually harmful since we don't want the behavior
to diverge due to the presence of a remap plugin.

This change removes the dodgy request URL rewriting code and uses
the same rewriting code that is used in the non-plugin case. This
is easier to understand and more reliable.