TS-1868: TSREMAP_*_STOP does not stop remap plugin evaluation chain
I have several remap plugins that I want to combine in a single
remap rule. According to the doc, I can chain them in remap.config
and if one remap plugin return TSREMAP_NO_REMAP_STOP or
TSREMAP_DID_REMAP_STOP, the evaluation chain will be broken (no
further plugin down the chain gets to run).

However, I have found out that this is not the case. All remap
plugins in the chain will run regardless of plugins' return code.
The only time I see it breaks the chain is if a plugin specifies
that it wants to return a redirection response to the mapped url.