TS-1918: SSL hangs after origin handshake
When using SSL with a Netscaler origin, I'm seeing the handshake
complete but ATS never issuing the HTTP request. It seems that while
the handshake is in progress, the write to the origin is not
rescheduled, so ATS never notices that it can send the origin
request.

Conflicts:
	CHANGES

test/review: jpeach, igalic, zwoop
backport: igalic