promote TS-1912 and clone as TS-1918
as a side-note, we have noted that this is process is stupid.
We should clone as soon as we propose: If something goes wrong, we can
close that new issue as invalid, but one way or the other, we can track
votes in the NEW issue, not in the old one, which was the whole point of
cloning them in the first place.