TS-1825: TSPortDescriptorAccept does not use it's argument
TSPortDescriptorAccept never calls the acceptance continuation that
it takes as an argument. Fix TSPortDescriptorAccept so that it
actually uses the continuation from the caller. Fix the corresponding
unit test.

Refactor the SDK_API_TSNetVConn and SDK_API_TSPortDescriptor tests
and update the them so that them actually verify the connection on
the server side. Generalize them to not use any global variables.
over