TS-1919 Eliminate CacheLookupHttpConfig
This breaks compatibility with the cache clustering, so all nodes
in a cluster would have to be upgraded at the same time. This
eliminates an ugly passing of certain configurations through
the cluster channels into the cache core.

In addition, I think this can open up the possibility to make
some cache related configurations overridable. That should be
done as a different commit though.