TS-1946: reduce the verbosity of SSL handshake errors
Update SSL error logging routines so that we can log the SSL error
stack at debug level as well as error level. Update SSL error logging
to log the location of the caller, rather than the location of the
logging function.

Drop SSL server handshake logs to debug level so that they don't
overwhelm the diagnostics log. Handshake errors can be very common
and are not generally interesting.