attempt to fix links in admin guide
I think what we'll end up doing here will simply be a collection of all
our links, thematically ordered, in a single file and include them.

Right now, this still doesn't work out perfectly, because it depends on
multiviews (or some form of negotiation). But otherwise, links to
*.config work.
Deeplinks into records.config don't work yet, because that file's a mess