formatting: replace defintions w/ field-lists
in records.config I have started to replace defintions with field-lists.
This seems to yield a much smoother look-and feel of the output.
Further, I am adding a ref, in the form of:

    .. _foo:

before the start of of such a :foo: -list.
It's not yet clear to me whether these anchors are only generated if
they are actually referenced. I guess that's to be seen.

n.b.: Right now, our style in particular, or sphinx in general,
generates those field-lists as tables, and adds

	<tbody valign="top">

This makes sure that the header of the next section are directly stuck
to the bottom of the previous defition. That needs to be fixed, and I
think we'll fix it in the CSS.