TS-1953: remove check_ts_version() from examples
remove the boilerplate check_ts_version() function call from all
examples. This function is nice in a specific example, but it does not
make sense to have it in every single example plugin.

Since the code is compiled, in most cases that's too late to catch
issues with API versions, so it doesn't make sense to begin with.
In order to enable Plugin Developers to do such checks, and #ifdef
around API issues, we'd have to expose the TS_VERSION_* contants.

n.b.: We might want add one example which shows off how to use the
TSTrafficServerGetVersion() API, but we have that in the man page now
anyway.