TS-1823: add line continuation to remap.config
Add line continuation to remap.config using a backslash at the end
of a line as the continuation character.

The handling for continuation was made in tokLine() by adding an
optional third char parameter (`cont') which is the continuation
character. This should make it usable in other contexts outside of
remap.config also, for example other config file parsers.

This implementation is not very intelligent, as it only checks for the
backslash immediately preceding the newline and does not handle any
whitespace. The backslash and newline are converted into spaces, and
the next line is appended.

Example:

.definefilter foo \
  @action=allow \
  @src_ip=127.0.0.1

is read and parsed in UrlRewrite::BuildTable() as:

.definefilter foo     @action=allow     @src_ip=127.0.0.1