TS-1597: Document remap.config filters
Add documentation and examples for the
.definefilter, .activatefilter, and .deactivatefilter
directives in remap.config.

These docs are specifically restricted to these particular names
because (1) we should standardize on one variant for each of these
directives to avoid end-user confusion and (2) I personally believe
these variants to be the clearest and most understandable ("defflt",
"useflt", "deuseflt", etc., are confusing or awkward).

The .deletefilter directive is not documented, because I am not sure
when or why it would be used.