TS-2137: Use eventfd instread of pthread signal/wait in ATS
pthread_cond_signal/wait is used in several places in ATS, including but
not limited:
1) Logging system.
2) ProtectedQueue in event system.
3) RecProcess in stats system.

As we known, pthread_cond_signal() need to take lock, it'll cause
more context switch than eventfd.

In my testing:
1) Using _signal()/_wait() pair, eventfd is about 5 times faster than
   pthread cond.

2) Using _signal()/_timedwait() pair, eventfd is about 2 times faster
   than pthread cond.

== NOTE ==
pthread_cond_signal/wait is also used by AIO module, but we can't
simply replace it with eventfd, as AIO code use the mutex to protect
other stuff. And I found we can't replace it in ProtectedQueue directly,
there are some prepare work to do.

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>