TS-2143 Remove alarm email address support (and config).
This removes proxy.config.admin.admin_user, which was used for
other things in the past, but now only for an email address
passed to the alarm script.

The alarm scripts now only get two arguments (and always 2):

    A description of the alarm
    A number identifying the alarm (e.g. "14")

In addition, this script cleans up some old code that was
customized for an old Inktomi customer. This is obviously
no longer something we need, or should support.

The "Upgrade to v4.0.0" Wiki will be updated with these details.