TS-2148: handle_cache_operation_on_forward_server_response ignores value of api_server_response_no_store
In the function HttpTransact::handle_cache_operation_on_forward_server_response
there is a line that checks the value of s->api_server_response_no_store
(line 4162 in HttpTransact.cc). This value will never be true,
however, because is_response_cacheable is called before this check
(on line 3934) and sets this value to false (on line 6104).