TS-2163 Remove obsolete WDA and ACC alarms feature
This was done for an Inktomi specific customer plugin, and no longer applies
to the project. If we later decide it necessary, we should add new APIs which
let plugins register new alarms with the system for appropriate handling.

Reading the code, it looks to me like the Alarms code was intended to be
replaced as well, so this could be part of a code refactoring/replacement.