TS-2168 Clean up the verbose comments
There's not much purpose to repeat the same thing 4 times :). This
also fixes a potential problem with debug_filter, since it currently
has two bit fields in use, we should allow for values 0-3, right?

There's still a huge problem with this code as it is now, because
it would make it impossible to implement features that does
consistency checks on records.config (see TS-1882).