TS-2212 Implement <fsiz> tag for HTTP
The current docs explains it as

``fsiz``
    The size of the file (*n* bytes) as seen by the origin server.

However, there's nothing that implements this. This code will
try its best to find a body size, based on origin response
headers such as Content-Range and bytes read.

It's somewhat unfortunate that at the time of logging, the CacheVC
is no longer evailable, and all headers are already modified such
that they reflect client's e.g. Range: bytes=  header.

This patch implements the logic around using the Range-response
headers when available, and the internal (full) byte counters
when not.

Finally, this also generalizes the "atoi" functionality in lib/ts
(ParseRules) such that the prototypes with and without a string
length are available for all int types. This was needed in order
to do an efficient and safe atoi on the not-null terminated string
to a 64-bit int.