TS-1988 cquuc and cquup can have no values
When there is no remap phase, the unmapped URL does not get
populated. This means that cquuc and cquup can have no value,
so a "-" is logged. This happens, for example, with server
intercept plugins that do not pass through a remap phase.

This patch falls back on the original client URL in these
situations.