TS-2245 This adds a '2' config state to the ignore mismatch configs
This is completley backwards compatible, but adds a new feature which
allows a much more sane behavior for these configs:

  proxy.config.http.cache.ignore_accept_mismatch
  proxy.config.http.cache.ignore_accept_language_mismatch
  proxy.config.http.cache.ignore_accept_encoding_mismatch
  proxy.config.http.cache.ignore_accept_charset_mismatch

Setting any of these to the value '2' will now ignore the
respective client request header, but *only* if the cached
response does not have a Vary: header.

The old documentation seemed to imply that this would be
the case, but it was never so. Setting any of these to "1"
would simply ignore whatever the Vary: header says.