Make targets for creating releases+release candidates
What this does
==============

* remove the check for .svn: No longer necessary since we moved to git.

* for the `release` target we simply reuse asf-dist-sign, and
additionally create a signed tag of the same name.

* for the `rel-candidate` target, we add a variable `RC` to Makefile.am
with default value of 0. This can be overriden from outside by calling

    make release-candidate RC=14

NOTE: tags are created with `-f`. If a tag of the same name existed on
your system, it will be overwritten. This is considered a feature.

What this does *not* do
=======================

We are not pushing any tags upstream, or uploading any tarballs
anywhere. This is in the responsibility of the Release Manager.

Likewise, we do *not* bump the version in configure.ac