reliably recreate the same tar-ball from the same content
if the content of the tree is the same, we now create the same
tarballs with the same and sha1sums. We do this, by fixing the mtime of
all objects in the tarball to that of our ./configure.ac

In order for the content to be the same, we no longer rely on simply
calling $(MAKE) asf-dist-sign with a different version, but rather have
special targets, which simply inject the $(RC) variable.

These targets are entirely convinience wrappers for Release managers.

NOTE the GPG signature will always different, because it's a cryptographic
signature. In order to release 4.2.16-rc7 as 4.2.16, we can either call

    make release

or, simply rename the tar-ball. The content of the tar-ball will be the
same, regardless.