TS-2361: improve regex_remap plugin loading semantics
The location of the regex_remap configuration is not well-defined
if it is a relative path. Either it will be found in the current
directory (which is undefined), or a hard-coded global directory
will be used.

This change always uses the absolute path if an absolute path is
provided. If a relative path is provided, then we load that relative
to the configured Traffic Server configuration directory.