TS-2401: Remove global system_log_dir
In the 2 places that actually look at this, we implement the
predictable semantics of using proxy.config.log.logfile_dir (relative
to $PREFIX) if it is specified. Otherwise we use the compile-time
log directory.  In either case, we never fall back to different
directories. We either use the one specified by the operator or
fail loudly and suddenly.