TS-2479: Don't output orphan log after failover sucessfully
Don't output orphan log after failover successfully, BTW fix a potential
issue about LogBuffer->m_references.

==NOTE==:
Some logs maybe pending in the buffer of network thread when log server
is down, in which case, the pending logs will be outputed to orphan file,
and them will not be sent to the failover host.

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>