Add a SSL certificate matching integration test
Various fixes the the tsqa harness to improve output legibility.
Crank up the configuration sync frequency so that we don't need to
wait so long for configuration changes.

Add a new integration test to ensure that SSL certificate matching
works correctly for wildcards, SNI names, IP addresses and address/port
combinations. Still missing is the default match; for that I need
a portable way to determine a local IP address.