TS-2519: make using RECP_NULL a compilation error
RECP_NULL can still be useful to represent the case when a record
does not have a persistence type. It might be a configuration record,
for example. However, stats records shold never be registered as
RECP_NULL, because that's ambiguous as to whether it should be
persisted This change add some template helpers to ensure that any
attempt to register a RECP_NULL stat dies in a horrible shower of
compiler errors.