TS-2239: support the ALPN TLS extension
Add support the for TLS Application Layer Protocol Negotiation
extension. This uses the same infrastructure as NPN, so a consistent
protocol advertisement is created. Any plugin that registers to
accept a named protocol will automatically gain NPN and ALPN support.

This requires a very recent OpenSSL, probably 1.2 or something like
that.