TS-2590: translate documentation into Japanese
We need 'sphinx-intl' and sphinx 1.2 or higher.
- sphinx-intl: https://pypi.python.org/pypi/sphinx-intl

The build commands are following:
  $ cd doc
  $ sphinx-intl build -p locale/ -l ja
  $ sphinx-build -b html -D language=ja . ./build

On the ReadTheDocs, we need only 'polib'. This settings has already
been proposed and merged in the apache/trafficserver #13.
But this works only in the ReadTheDocs environment. So, using sphinx-intl
is easier.

Authors: Masaori Koshiba <masaori335@gmail.com>
         Masakazu Kitajo <m4sk17@gmail.com>
         syucream <syucream1031@gmail.com>