TS-2569: set the default SSL options correctly
We discovered that the proxy.config.ssl.server.honor_cipher_order=1
setting was not working correctly. After investigating it was
determined that if you do not have a dest_ip=* in the ssl_multicert.config
file then the server cipher order setting will not be honored.  The
proposed fix (which works) is to initialize the default context with
the necessary SSL options.

(cherry picked from commit 963982e432a6fa5ef0f1968904c75571a3f6befb)

Conflicts:
	CHANGES
	iocore/net/SSLUtils.cc