TS-2630: Add lib/ts/apidefs.h to place common types
The apidefs.h file is used to place common types that can be shared
between internal code and TS APIs.

It's hard to judge what types in api/ts/ts.h should be migrated to
lib/ts/apidefs.h, so this patch is just a beginning ...

Why I named it 'apidefs.h' instead of 'defs.h' and put it into 'lib/ts'?
 1) Firstly, there is a 'ink_defs.h' file in 'lib/ts', then 'defs.h' seems
    too similar with it. So named it 'apidefs.h' would be more clearly.

 2) Since 'apidefs.h' will be shared not only by 'iocore' modules, puting it
    into 'lib/ts' would be more suitable than 'iocore/api' or any other else.

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>