TS-2431: Migrate Taobao SPDY plugin to ATS core
This path depends on FetchSM, so seriously speaking, the migration
to ATS core is finished *partially*. I had tried to remove the dependency
of FetchSM and create a specific VC for each http request in spdy sm, but I
found it was not so easy, so I give up temporary. Let me push this
patch to community before it's perfect enough, at least, this series
patches can statisfy TAOBAO's current demand.

With this patch, ATS supports SPDY(v3/v3.1), the SPDY can run with/without
SSL:
 1) SPDY without SSL:
    It will share the same port number with HTTP, and ATS will recognize
    SPDY/HTTP by detecting the first byte of client request.

 2) SPDY with SSL:
    When running SPDY with SSL, the code needs OpenSSL(>=1.01), if this
    version of OpenSSL is installed in user-defined DIR, you may need to
    use '--with-openssl=<dir>' option to tell ATS where to search.

Signed-off-by: Yunkai Zhang <qiushu.zyk@taobao.com>